Modify progressBar component - Modify Login screen
