+++ components/ProgressBar/index.tsx
@@ -1,17 +1,32 @@
+import { useState, useEffect } from "react";
 import styles from "./styles.module.css";
 
 export interface IProgressBar {
-  value: number;
-  max: number;
   message: string;
 }
 
-function ProgressBar({ value, max, message }: IProgressBar) {
+function ProgressBar({ message }: IProgressBar) {
+  const [value, setValue] = useState(0);
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setValue((oldValue) => {
+        const newValue = oldValue + 20;
+
+        if (newValue === 100) {
+          clearInterval(interval);
+        }
+
+        return newValue;
+      });
+    }, 1000);
+  }, []);
+
   return (
-    <div>
-      <progress value={value} max={max} className={styles.progressBar} />
+    <>
+      <progress value={value} max={100} className={styles.progressBar} />
       <p>{message}</p>
-    </div>
+    </>
   );
 }
